chore(pagination): rename table.paginator -> structuredElementPaginator.js
