+++ src/node/elements/structuredElementPaginator.js
